requirements.txt
androidstorage4kivy==0.1.1
certifi==2024.12.14
chardet==5.2.0
charset-normalizer==3.4.1
Cython==3.0.11
docutils==0.21.2
filetype==1.2.0
idna==3.10
Kivy==2.3.1
kivy-deps.angle==0.4.0
kivy-deps.glew==0.3.1
Kivy-Garden==0.1.5
kivy_deps.sdl2==0.8.0
pillow==11.1.0
plyer==2.1.0
Pygments==2.19.1
pyjnius==1.6.1
pypiwin32==223
pywin32==308
reportlab==4.3.0
requests==2.32.3
urllib3==2.3.0
requests>=2.25.0
packaging>=21.0